Build und publish the generated static files as a branch on GitHub
Adding a manual GitHub workflow to create a new release from the
'main' branch. Creates a release/... branch with the static files
generated by Vite. This way we can serve them through a CDN.

We adapt the Vite/Roillup config to not use hashes in the generated
file names, so the file name is fixed and does not change between
versions.